refine the timeskipping unit test
